requirements.txt
requests==2.31.0
pandas==2.2.0
fastapi==0.109.0
uvicorn==0.27.0
groq>=0.5.0
httpx>=0.27.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.5.3
